reverted move of AsyncMutex to Utils
