+++ src/Microsoft.TemplateEngine.Utils/PublicAPI.Unshipped.txt
@@ -1,8 +1,4 @@
-﻿Microsoft.TemplateEngine.Utils.AsyncMutex
-Microsoft.TemplateEngine.Utils.AsyncMutex.Dispose() -> void
-Microsoft.TemplateEngine.Utils.AsyncMutex.IsLocked.get -> bool
-Microsoft.TemplateEngine.Utils.IFileSystemInfoExtensions
-static Microsoft.TemplateEngine.Utils.AsyncMutex.WaitAsync(string! mutexName, System.Threading.CancellationToken token) -> System.Threading.Tasks.Task<Microsoft.TemplateEngine.Utils.AsyncMutex!>!
+﻿Microsoft.TemplateEngine.Utils.IFileSystemInfoExtensions
 static Microsoft.TemplateEngine.Utils.IFileSystemInfoExtensions.GetDisplayPath(this Microsoft.TemplateEngine.Abstractions.Mount.IFileSystemInfo! fileSystemInfo) -> string!
 static Microsoft.TemplateEngine.Utils.WellKnownSearchFilters.AuthorFilter(string? author) -> System.Func<Microsoft.TemplateEngine.Abstractions.ITemplateInfo!, Microsoft.TemplateEngine.Abstractions.TemplateFiltering.MatchInfo?>!
 static Microsoft.TemplateEngine.Utils.WellKnownSearchFilters.BaselineFilter(string? baselineName) -> System.Func<Microsoft.TemplateEngine.Abstractions.ITemplateInfo!, Microsoft.TemplateEngine.Abstractions.TemplateFiltering.MatchInfo?>!
